+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export type {
   MeshkitClient,
   MeshkitConfig,
   MeshkitInitOptions,
+  StoredObject,
   IpnsDuration,
   IpnsKey,
   IpnsKeyGenOptions,
@@ -14,8 +15,8 @@ export { MeshkitError } from './types.js';
 export { Meshkit } from './meshkit.js';
 
 export { createMeshkitClient } from './create-client.js';
-export { createFilOneClient } from './create-filone-client.js';
-export type { FilOneConfig } from './create-filone-client.js';
+export { createS3Client, createFilOneClient } from './create-filone-client.js';
+export type { S3StorageConfig, FilOneConfig } from './create-filone-client.js';
 
 export { IPNS_TTL_DEFAULT, IPNS_TTL_FAST } from './ipns/constants.js';
 export { extractCidFromPath, toIpfsPath, toIpnsPath } from './ipns/paths.js';